placate AVR and SAM compiler

+++ src/settings/i2c/CAT24I2CFlashSettingsStorage.cpp
@@ -78,7 +78,7 @@ void CAT24I2CFlashSettingsStorage::reset() {
 
 
 int CAT24I2CFlashSettingsStorage::readBytes(void* valueToSet, int numBytes) {
-    int read = _wire->requestFrom(_address, (size_t)numBytes, true);
+    int read = _wire->requestFrom((uint8_t)_address, (uint8_t)numBytes, (uint8_t)true);
     _currptr += read;
     if (read==numBytes)
         _wire->readBytes((uint8_t*)valueToSet, numBytes);
